merges v20.08 tag

Test: build

+++ src/core/NEON/kernels/arm_gemm/transforms/a64_block16_interleave4_8bit.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2018 ARM Limited.
+ * Copyright (c) 2017-2020 Arm Limited.
  *
  * SPDX-License-Identifier: MIT
  *
@@ -39,7 +39,7 @@ void TransformImpl<4, 16, false, 1, 1, false>::Transform(T *out, const T *in, in
     uint8_t zerobuff[16] = { 0 };
 
     for (int y=y0; y<ymax; y+=4) {
-        const uint8_t *inptr0 = inptr + y * ldin + k0;
+        const uint8_t *inptr0 = inptr + static_cast<intptr_t>(y) * ldin + k0;
         const uint8_t *inptr1 = inptr0 + ldin;
         const uint8_t *inptr2 = inptr1 + ldin;
         const uint8_t *inptr3 = inptr2 + ldin;
@@ -54,7 +54,6 @@ void TransformImpl<4, 16, false, 1, 1, false>::Transform(T *out, const T *in, in
             /* Cope with ragged cases by copying from a buffer of zeroes instead */
             if ((y + 3) >= ymax) {
                 switch ((y + 3) - ymax) {
-                    /* Everything falls through in here */
                     case 2:
                         inptr1 = zerobuff;
                         // fall through
@@ -92,7 +91,6 @@ void TransformImpl<4, 16, false, 1, 1, false>::Transform(T *out, const T *in, in
             /* Need to duplicate this here, in case we didn't run the main loop. */
             if ((y + 3) >= ymax) {
                 switch ((y + 3) - ymax) {
-                    /* Everything falls through in here */
                     case 2:
                         inptr1 = zerobuff;
                         // fall through
